Updated assign.fold() in plm_utils.r
foldid will be generated within the function and not be supplied to the function, since the function modified foldid anyway.
q(status=1) was exchanged for a stop() warning message.
check if all classes labels are in all folds only for stratified CV, for non-stratified/LOO-CV, check only that all classes are in the training fold.