Make the alpha-value for ENET customizable
The range in which the alpha is chosen for the enet algorithm was not customizable, but hard-coded between 0 and 1 (which makes a lot of sense). But, if someone wants to search a smaller window of alpha, it was impossible. Now, if alpha is supplied in the parameter set, the value from this list will be taken, and not the hardcoded defaults.