Bugfix for summarize.features
Summarize features did not work correctly before, since matching of higher order names was not performed correctly. For example, the features for the genus "g__Ruminococcus2" would be counted as well when summarizing the features of the genus "g__Ruminoccous" (without the "2").
Now, the feature names are split by the identifiers of the taxonomic rank and the matching is performed in the resulting table. The function will needs some checks of the feature names in the future.