Added Variance filter to filter.features
Added variance filter to the filter.features function. Additionally, fixed some bugs (cum.abundance method was misspelled in inital checks), fixed some messaging, and added a check for the number of features that would be retained after filtering.

Additionally, added import of the function rowVars in the import_packages file.

Furthermore, adjusted the documentation of filter.features to match the new function.