Fix errors introduced by new roxygen version
the new roxygen version tried to be helpful and to detect S3method in functions with a dot. For example, filter.features was detected as S3method(filter, features) and exported as such in the NAMESPACE. Therefore, the package could not be installed properly.

Fixed now for all four functions that had this problem (`filter.features`, `filter.label`, `summarize.features`, and `select.samples`)