Bugfix for add.meta.pred
Addition of meta-variables to the feature table was implemented incorrectly so far. Instead of adding the meta-variables one after the other, only the last meta-variable ended up in the returned SIAMCAT object. Corrected by taking the assignment of the updated feature table outside of the for loop for the different meta-variables.